+++ Demos/MATLAB/Model2DGeneratorDemo.m
@@ -10,16 +10,18 @@
 close all;clc;clear;
 % adding paths
 fsep = '/';
-addpath('compiled'); addpath('supplem'); 
+Path1 = sprintf(['..' fsep '..' fsep 'Wrappers/MATLAB/compiled'], 1i);
+Path2 = sprintf(['..' fsep '..' fsep 'Wrappers/MATLAB/supplem'], 1i);
+addpath('Path1'); addpath('Path2'); 
 
-ModelNo = 4; % Select a model from Phantom2DLibrary.dat
+ModelNo = 13; % Select a model from Phantom2DLibrary.dat
 % Define phantom dimensions
-N = 512; % x-y size (squared image)
+N = 1024; % x-y size (squared image)
 
 % Generate 2D phantom:
 curDir   = pwd;
 mainDir  = fileparts(curDir);
-pathtoLibrary = sprintf([fsep '..' fsep 'PhantomLibrary' fsep 'models' fsep 'Phantom2DLibrary.dat'], 1i);
+pathtoLibrary = sprintf(['..' fsep '..' fsep 'PhantomLibrary' fsep 'models' fsep 'Phantom2DLibrary.dat'], 1i);
 pathTP = strcat(mainDir, pathtoLibrary); % path to TomoPhantom parameters file
 [G] = TomoP2DModel(ModelNo,N,pathTP); 
 figure; imagesc(G, [0 1]); daspect([1 1 1]); colormap hot;