+++ CHANGES.md
@@ -1,5 +1,11 @@
 # ChangeLog
 
+## v1.4.3
+* Artifacts simulation module has been modified. It is based now on specifying the dictionaries where artifact types are
+described. Stipes can be simulated to be partial and with variable intensity.
+* 2D Model 15 added - DLS phantom
+
+
 ## v1.4.2
 * Jupiter Notebook demo added which uses deep learning algorithms to employ TomoPhantom for data generation